# Conflicts:
#	composer.json
#	config/app.php

+++ packages/Webkul/Product/src/Listeners/ProductFlat.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Webkul\Attribute\Repositories\AttributeRepository;
 use Webkul\Attribute\Repositories\AttributeOptionRepository;
+use Webkul\Product\Helpers\ProductType;
 use Webkul\Product\Repositories\ProductFlatRepository;
 use Webkul\Product\Repositories\ProductAttributeValueRepository;
 use Webkul\Product\Models\ProductAttributeValue;
@@ -148,7 +149,7 @@ class ProductFlat
     {
         $this->createFlat($product);
 
-        if ($product->type == 'configurable') {
+        if (ProductType::hasVariants($product->type)) {
             foreach ($product->variants()->get() as $variant) {
                 $this->createFlat($variant, $product);
             }
@@ -229,11 +230,7 @@ class ProductFlat
                             }
                         }
 
-                        if ($product->type == 'configurable' && $attribute->code == 'price') {
-                            $productFlat->{$attribute->code} = app('Webkul\Product\Helpers\Price')->getVariantMinPrice($product);
-                        } else {
-                            $productFlat->{$attribute->code} = $productAttributeValue[ProductAttributeValue::$attributeTypeFields[$attribute->type]];
-                        }
+                        $productFlat->{$attribute->code} = $productAttributeValue[ProductAttributeValue::$attributeTypeFields[$attribute->type]];
 
                         if ($attribute->type == 'select') {
                             $attributeOption = $this->attributeOptionRepository->find($product->{$attribute->code});
@@ -270,6 +267,10 @@ class ProductFlat
 
                     $productFlat->updated_at = $product->updated_at;
 
+                    $productFlat->min_price = $product->getTypeInstance()->getMinimalPrice();
+
+                    $productFlat->max_price = $product->getTypeInstance()->getMaximamPrice();
+
                     if ($parentProduct) {
                         $parentProductFlat = $this->productFlatRepository->findOneWhere([
                                 'product_id' => $parentProduct->id,
@@ -277,21 +278,24 @@ class ProductFlat
                                 'locale' => $locale->code
                             ]);
 
-                        if ($parentProductFlat) {
+                        if ($parentProductFlat)
                             $productFlat->parent_id = $parentProductFlat->id;
-                        }
                     }
 
                     $productFlat->save();
                 }
             } else {
-                $productFlat = $this->productFlatRepository->findOneWhere([
-                    'product_id' => $product->id,
-                    'channel' => $channel->code,
-                ]);
+                $route = request()->route() ? request()->route()->getName() : "";
 
-                if ($productFlat) {
-                    $this->productFlatRepository->delete($productFlat->id);
+                if ($route == 'admin.catalog.products.update') {
+                    $productFlat = $this->productFlatRepository->findOneWhere([
+                        'product_id' => $product->id,
+                        'channel' => $channel->code,
+                    ]);
+
+                    if ($productFlat) {
+                        $this->productFlatRepository->delete($productFlat->id);
+                    }
                 }
             }
         }
